[BOLT] Re-type BinaryContext::HugePageSize to uint64_t
... so that it matches that of BinaryContext::PageAlign.

+++ bolt/include/bolt/Core/BinaryContext.h
@@ -648,7 +648,7 @@ public:
   static constexpr unsigned RegularPageSizeLoongArch = 0x4000;
 
   /// Huge page size to use.
-  static constexpr unsigned HugePageSize = 0x200000;
+  static constexpr uint64_t HugePageSize = 0x200000;
 
   /// Addresses reserved for kernel on x86_64 start at this location.
   static constexpr uint64_t KernelStartX86_64 = 0xFFFF'FFFF'8000'0000;
